refactor:   domain - enforce use of make() to create movable board or number; use 'auto' keyword  when is secure to use

+++ src/2048/domain/score.hpp
@@ -3,6 +3,6 @@
 
 #include <cstdint>
 
-using Score = int16_t;
+using Score = int32_t;
 
 #endif
